added interupt handling

+++ src/modules/aob-signature-ptr.cpp
@@ -6,14 +6,14 @@
 #include "aob-signature-ptr.h"
 #include "memory-span.h"
 
-AOBSignaturePtr::AOBSignaturePtr(std::string pattern, uintptr_t begin, MemorySpan scan_span)
+AOBSignaturePtr::AOBSignaturePtr(std::string pattern, int begin, MemorySpan scan_span)
 {
     this->signature = new AOBSignature(pattern);
     this->begin = begin;
     this->scan_spans.push_back(scan_span);
 }
 
-AOBSignaturePtr::AOBSignaturePtr(std::string pattern, uintptr_t begin, const std::vector<MemorySpan>& scan_spans)
+AOBSignaturePtr::AOBSignaturePtr(std::string pattern, int begin, const std::vector<MemorySpan>& scan_spans)
 {
     this->signature = new AOBSignature(pattern);
     this->begin = begin;
@@ -25,7 +25,7 @@ const AOBSignature& AOBSignaturePtr::getSignature() const
     return *(this->signature);
 }
 
-uintptr_t AOBSignaturePtr::getBegin() const
+int AOBSignaturePtr::getBegin() const
 {
     return this->begin;
 }
